Remove rust-toolchain file (#99)
Architecture specific toolchains are not allowed in this file.
We can assume that stable is the default toolchain and should not
dictate with which toolchain the user builds the client. Also this
breaks under macOS.

We are only interested with which toolchain the runtime is build.
For this we should set the WASM_BUILD_TOOLCHAIN variable.